refactor(martech): move url parameter tracking functions to new file
